requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
dnspython==1.16.0
eventlet==0.31.0
greenlet==1.1.0
idna==2.10
pkg-resources==0.0.0
python-engineio==4.2.0
python-socketio==5.3.0
requests==2.25.1
six==1.16.0
urllib3==1.26.4
